extract api hook to separate file
