+++ application/ui/src/constants/shared-types.ts
@@ -78,3 +78,5 @@ export type SourceConfigPayload = Exclude<SourceConfig, DisconnectedSourceConfig
 
 export type AnnotationDTO = components['schemas']['DatasetItemAnnotation-Input'];
 export type DatasetItemAnnotationStatus = components['schemas']['DatasetItemAnnotationStatus'];
+
+export type AnnotatedVideoFrame = components['schemas']['AnnotatedVideoFrame'];
